Final version

+++ AccountingSystem/Services/SalesCategoryData.cs
@@ -3,9 +3,9 @@
     public class SalesCategoryData
     {
         public string Category { get; set; }
-        public double TotalSales { get; set; }
+        public float TotalSales { get; set; }
         public double TotalProfit { get; set; }
-        public int SalesCount { get; set; } // Кількість продажів
+        public int SalesCount { get; set; }
 
 
     }
